python/pyproject.toml
[tool.poetry]
name = "performance"
version = "0.1.0"
description = ""
authors = ["Rodrigo F Martins <martins.drg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
tqdm = "^4.66.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
